fix: Move classes around to prevent circular imports
